[jaypedia/airbnb#29] Refactor: Api별로 Axios폴더 분리
