+++ FE/src/apis/searchBar.ts
@@ -1,0 +1,10 @@
+import axios from '@/apis/core/searchBar';
+
+export default {
+  getRangePriceData() {
+    return axios({
+      url: '/prices',
+      method: 'get',
+    });
+  },
+};
